% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVSO.R
\name{CIVSO}
\alias{CIVSO}
\title{Main CIVSO Estimation Function}
\usage{
CIVSO(
  betaX,
  betaY,
  seX,
  seY,
  ld_score,
  n_snp,
  n_x,
  n_y,
  overlap_prop,
  method = "diagonal_WLS",
  param_alpha = 0.05
)
}
\description{
Main CIVSO Estimation Function
}
