% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_civso.R
\name{.civso_engine}
\alias{.civso_engine}
\title{Internal Engine: Core CIVSO Estimator (Single Iteration)}
\usage{
.civso_engine(
  idx_subset = NULL,
  betaX,
  betaY,
  seX,
  seY,
  ld_score,
  n_snp,
  n_x,
  n_y,
  overlap_prop,
  covXY_theory,
  method = "diagonal",
  blocks = NULL
)
}
\description{
Internal Engine: Core CIVSO Estimator (Single Iteration)
}
\keyword{internal}
