% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.commutation_matrix}
\alias{.commutation_matrix}
\title{Internal Helper: Commutation Permutation Vector
Returns indices to permute columns for the Commutation Matrix K}
\usage{
.commutation_matrix(m)
}
\description{
Internal Helper: Commutation Permutation Vector
Returns indices to permute columns for the Commutation Matrix K
}
\keyword{internal}
