% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_full_moment_dMAR.R
\name{.estimate_block_gls_hxy}
\alias{.estimate_block_gls_hxy}
\title{Internal Engine: Full-Moment GLS Cross-Trait (Block-wise)}
\usage{
.estimate_block_gls_hxy(
  betaX,
  betaY,
  R,
  n_total,
  n_x,
  n_y,
  overlap_prop,
  covXY_theory,
  res_XX_init,
  res_YY_init,
  h_xy_init,
  v_xy_init
)
}
\description{
Internal Engine: Full-Moment GLS Cross-Trait (Block-wise)
}
\keyword{internal}
