% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_diagonal_dMAR.R
\name{.estimate_h2_ols}
\alias{.estimate_h2_ols}
\title{Internal Engine: Simple OLS Heritability (Unweighted)}
\usage{
.estimate_h2_ols(beta, se, ld_score, n_snp, n_samp)
}
\description{
Performs unweighted regression (w=1) for single-trait heritability.
Used as a baseline.
}
\keyword{internal}
