% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{.estimate_hxy_ols}
\alias{.estimate_hxy_ols}
\title{Internal Engine: Simple OLS (Unweighted)}
\usage{
.estimate_hxy_ols(
  betaX,
  betaY,
  ld_score,
  covXY_theory,
  n_snp,
  n_x,
  n_y,
  overlap_prop
)
}
\description{
Performs unweighted regression (w=1).
Used as a baseline to demonstrate the value of WLS.
}
\keyword{internal}
